pyproject.toml
[project]
name = "job-finder"
version = "0.1.0"
description = "Job search automation tool"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.54.0",
    "pandas>=2.3.3",
    "pymupdf>=1.25.0",
    "python-docx>=1.1.0",
    "python-jobspy>=1.1.82",
    "google-genai>=1.64.0",
    "rapidfuzz>=3.14.3",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0",
    "bcrypt>=4.0",
    "apscheduler>=3.10",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-mock>=3.15.1",
    "hypothesis>=6.0.0",
]
